web/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["juyony <wnsdyd1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"
fastapi = {extras = ["all"], version = "^0.110.1"}
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
pandas = "^2.2.2"
requests = "^2.31.0"
konlpy = "^0.6.0"
mecab-python3 = "^1.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
